backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Dance==7.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==3.0.1
SQLAlchemy==2.0.23
anthropic==0.8.0
openai>=1.30.0
sendgrid==6.11.0
PyPDF2==3.0.1
requests==2.31.0
psycopg2-binary==2.9.9
PyJWT==2.10.1
pgvector==0.2.4
alembic==1.13.1
twilio==8.13.0
numpy>=1.24.0
langgraph>=0.2.0
langchain-core>=0.3.0
